+++ CHANGELOG.rst
@@ -11,6 +11,11 @@ adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 `Unreleased`_
 -------------
 
+Nothing yet
+
+`0.2.0`_ - 2019-09-04
+---------------------
+
 Added
 ~~~~~
 
@@ -256,7 +261,8 @@ Changed
    ``pyroomacoustics.datasets.timit``
 
 
-.. _Unreleased: https://github.com/LCAV/pyroomacoustics/compare/v0.1.23...HEAD
+.. _Unreleased: https://github.com/LCAV/pyroomacoustics/compare/v0.2.0...HEAD
+.. _0.2.0: https://github.com/LCAV/pyroomacoustics/compare/v0.1.23...v0.2.0
 .. _0.1.23: https://github.com/LCAV/pyroomacoustics/compare/v0.1.22...v0.1.23
 .. _0.1.22: https://github.com/LCAV/pyroomacoustics/compare/v0.1.21...v0.1.22
 .. _0.1.21: https://github.com/LCAV/pyroomacoustics/compare/v0.1.20...v0.1.21